rename trickFunctions.js to cardFunctions.js
