+++ src/utils/deckFunctions.js
